Standardize errors and add missing body parameter to put method
It wasn't possible to set body using put method, it is now possible.
Calling a client method with wrong parameter's type will throw a TypeError instead of an Error.
All Web Service authentication errors will now reject the promise with an AuthenticationError instead of an Error.
All Web Service request errors will now reject the promise with an RequestError instead of an Error.
404 request errors are now handled.